backend/requirements.txt
flask
flask_cors
psycopg2-binary
python-dotenv